backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
cryptography==41.0.7
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0
groq==0.4.1
httpx==0.27.0
google-api-python-client==2.149.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
